bonfire-quest-game/requirements.txt
python-dotenv>=1.0.1
fastapi>=0.115
uvicorn[standard]>=0.30
httpx>=0.27
